Moved collision checking into a plugin
