+++ exotations/collision_scene_fcl/exotica_plugins.xml
@@ -1,0 +1,5 @@
+<library path="lib/libcollision_scene_fcl">
+  <class name="exotica/CollisionSceneFCL" type="exotica::CollisionSceneFCL" base_class_type="exotica::CollisionScene">
+    <description>FCL collision scene</description>
+  </class>
+</library>
