+++ exotations/collision_scene_fcl/package.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>collision_scene_fcl</name>
+  <version>0.0.0</version>
+  <description>The collision checking using the FCL library.</description>
+  <maintainer email="vivan@ed.ac.uk">Vladimir Ivan</maintainer>
+
+  <license>BSD</license>
+
+  <buildtool_depend>catkin</buildtool_depend>
+  <depend>exotica</depend>
+  <depend>geometric_shapes</depend>
+  <depend>fcl</depend>
+
+  <export>
+    <exotica plugin="${prefix}/exotica_plugins.xml" />
+  </export>
+</package>
