+++ exotica/include/exotica/CollisionScene.h
@@ -1,0 +1,137 @@
+#ifndef COLLISIONSCENE_H
+#define COLLISIONSCENE_H
+
+#include <exotica/Tools.h>
+#include <exotica/Object.h>
+#include <exotica/Factory.h>
+#include <string>
+#include <sstream>
+#include <Eigen/Dense>
+#include <exotica/KinematicElement.h>
+#include <unordered_map>
+#include <unordered_set>
+
+#define REGISTER_COLLISION_SCENE_TYPE(TYPE, DERIV) EXOTICA_REGISTER(exotica::CollisionScene, TYPE, DERIV)
+namespace exotica
+{
+
+class AllowedCollisionMatrix
+{
+public:
+    AllowedCollisionMatrix() {}
+    AllowedCollisionMatrix(const AllowedCollisionMatrix& acm) {entries_ = acm.entries_;}
+    inline void clear() {entries_.clear();}
+    inline bool hasEntry(const std::string& name) const {return entries_.find(name)==entries_.end();}
+    inline void setEntry(const std::string& name1, const std::string& name2) {entries_[name1].insert(name2);}
+    inline void getAllEntryNames(std::vector<std::string>& names) const
+    {
+        names.clear();
+        for(auto& it : entries_)
+        {
+            names.push_back(it.first);
+        }
+    }
+    inline bool getAllowedCollision(const std::string& name1, const std::string& name2) const
+    {
+        auto it = entries_.find(name1);
+        if(it==entries_.end()) return true;
+        return it->second.find(name2)==it->second.end();
+    }
+private:
+    std::unordered_map<std::string, std::unordered_set<std::string>> entries_;
+};
+
+struct CollisionProxy
+{
+    CollisionProxy() : e1(nullptr), e2(nullptr), distance(0) {}
+    KinematicElement* e1;
+    KinematicElement* e2;
+    Eigen::Vector3d contact1;
+    Eigen::Vector3d normal1;
+    Eigen::Vector3d contact2;
+    Eigen::Vector3d normal2;
+    double distance;
+
+    inline std::string print() const
+    {
+        std::stringstream ss;
+        if(e1 && e2)
+        {
+            ss<<"Proxy: '"<<e1->Segment.getName()<<"' - '"<<e2->Segment.getName()<<"', c1: "<<contact1.transpose()<<" c1: "<<contact2.transpose()<<" d: "<<distance;
+        }
+        else
+        {
+            ss<<"Proxy (empty)";
+        }
+        return ss.str();
+    }
+};
+
+/// The class of collision scene
+class CollisionScene : public Uncopyable
+{
+public:
+
+    CollisionScene() {}
+
+    /**
+       * \brief Destructor
+       */
+    virtual ~CollisionScene() {}
+
+    /**
+       * \brief Check if the whole robot is valid (collision only).
+       * @param self Indicate if self collision check is required.
+       * @return True, if the state is collision free..
+       */
+    virtual bool isStateValid(bool self = true) = 0;
+
+    ///
+    /// \brief Computes collision distances.
+    /// \param self Indicate if self collision check is required.
+    /// \param computePenetrationDepth If set to true, accurate penetration depth is computed.
+    /// \return Collision proximity objectects for all colliding pairs of objects.
+    ///
+    virtual std::vector<CollisionProxy> getCollisionDistance(bool self, bool computePenetrationDepth = true) {throw_pretty("Not implemented!");}
+
+    /**
+       * @brief      Gets the collision world links.
+       * @return     The collision world links.
+       */
+    virtual std::vector<std::string> getCollisionWorldLinks() = 0;
+
+    /**
+       * @brief      Gets the collision robot links.
+       * @return     The collision robot links.
+       */
+    virtual std::vector<std::string> getCollisionRobotLinks() = 0;
+
+    virtual Eigen::Vector3d getTranslation(const std::string & name) = 0;
+
+    inline void setACM(const AllowedCollisionMatrix& acm)
+    {
+        acm_ = acm;
+    }
+
+    ///
+    /// \brief Creates the collision scene from kinematic elements.
+    /// \param objects Vector kinematic element pointers of collision objects.
+    ///
+    virtual void updateCollisionObjects(const std::map<std::string, std::shared_ptr<KinematicElement>>& objects) = 0;
+
+    ///
+    /// \brief Updates collision object transformations from the kinematic tree.
+    ///
+    virtual void updateCollisionObjectTransforms() = 0;
+
+protected:
+
+    /// The allowed collisiom matrix
+    AllowedCollisionMatrix acm_;
+};
+
+typedef exotica::Factory<exotica::CollisionScene> CollisionScene_fac;
+typedef std::shared_ptr<CollisionScene> CollisionScene_ptr;
+}
+
+#endif // COLLISIONSCENE_H
