+++ exotica/include/exotica/Scene.h
@@ -38,215 +38,12 @@
 #include "exotica/KinematicTree.h"
 #include <exotica/Property.h>
 #include <exotica/SceneInitializer.h>
-
-//  For collision
-#include <moveit/collision_detection/world.h>
-#include <moveit/collision_detection/collision_world.h>
-#include <moveit/collision_detection_fcl/collision_common.h>
-#include <fcl/broadphase/broadphase.h>
-#include <fcl/narrowphase/narrowphase.h>
-#include <fcl/collision.h>
-#include <fcl/distance.h>
-#include <moveit/robot_model_loader/robot_model_loader.h>
+#include <exotica/CollisionScene.h>
 #include <moveit_msgs/PlanningScene.h>
 #include <moveit/planning_scene/planning_scene.h>
 
-#include <moveit_msgs/DisplayRobotState.h>
-#include <moveit_msgs/DisplayTrajectory.h>
-#include <moveit/robot_state/conversions.h>
-
-#include <geometric_shapes/mesh_operations.h>
-#include <geometric_shapes/shape_operations.h>
-#include <unordered_set>
-#include <unordered_map>
-
 namespace exotica
 {
-  class Scene;
-
-  class AllowedCollisionMatrix
-  {
-  public:
-      AllowedCollisionMatrix();
-      AllowedCollisionMatrix(const AllowedCollisionMatrix& acm);
-      void clear();
-      bool hasEntry(const std::string& name) const;
-      void setEntry(const std::string& name1, const std::string& name2);
-      void getAllEntryNames(std::vector<std::string>& names) const;
-      bool getAllowedCollision(const std::string& name1, const std::string& name2) const;
-  private:
-      std::unordered_map<std::string, std::unordered_set<std::string>> entries_;
-  };
-
-  struct CollisionProxy
-  {
-      CollisionProxy() : e1(nullptr), e2(nullptr), distance(0) {}
-      KinematicElement* e1;
-      KinematicElement* e2;
-      Eigen::Vector3d contact1;
-      Eigen::Vector3d normal1;
-      Eigen::Vector3d contact2;
-      Eigen::Vector3d normal2;
-      double distance;
-
-      std::string print() const
-      {
-          std::stringstream ss;
-          if(e1 && e2)
-          {
-              ss<<"Proxy: '"<<e1->Segment.getName()<<"' - '"<<e2->Segment.getName()<<"', c1: "<<contact1.transpose()<<" c1: "<<contact2.transpose()<<" d: "<<distance;
-          }
-          else
-          {
-              ss<<"Proxy (empty)";
-          }
-          return ss.str();
-      }
-  };
-/// The class of collision scene
-  class CollisionScene : public Uncopyable
-  {
-    public:
-
-      struct CollisionData
-      {
-          CollisionData(CollisionScene* scene) : Scene(scene), Done(false), Self(true) {}
-
-          fcl::CollisionRequest Request;
-          fcl::CollisionResult Result;
-          CollisionScene* Scene;
-          bool Done;
-          bool Self;
-      };
-
-      struct DistanceData
-      {
-          DistanceData(CollisionScene* scene) : Scene(scene), Self(true), Distance{1e300} {}
-
-          fcl::DistanceRequest Request;
-          fcl::DistanceResult Result;
-          CollisionScene* Scene;
-          std::vector<CollisionProxy> Proxies;
-          double Distance;
-          bool Self;
-      };
-
-      struct ContactData
-      {
-          ContactData(CollisionScene* scene) : Scene(scene), Self(true), Distance{1e300} {}
-
-          fcl::DistanceRequest DistanceRequest;
-          fcl::DistanceResult DistanceResult;
-          fcl::CollisionRequest Request;
-          fcl::CollisionResult Result;
-          CollisionScene* Scene;
-          std::vector<CollisionProxy> Proxies;
-          double Distance;
-          bool Self;
-      };
-
-      CollisionScene(const std::string& root_name);
-
-      /**
-       * \brief Destructor
-       */
-      virtual ~CollisionScene();
-
-      static bool isAllowedToCollide(fcl::CollisionObject* o1, fcl::CollisionObject* o2, bool self, CollisionScene* scene);
-      static bool collisionCallback(fcl::CollisionObject* o1, fcl::CollisionObject* o2, void* data);
-      static bool collisionCallbackDistance(fcl::CollisionObject* o1, fcl::CollisionObject* o2, void* data, double& dist);
-      static bool collisionCallbackContacts(fcl::CollisionObject* o1, fcl::CollisionObject* o2, void* data, double& dist);
-
-      /**
-       * \brief Check if the whole robot is valid (collision only).
-       * @param self Indicate if self collision check is required.
-       * @return True, if the state is collision free..
-       */
-      bool isStateValid(bool self = true);
-
-      ///
-      /// \brief Computes collision distances.
-      /// \param self Indicate if self collision check is required.
-      /// \param computePenetrationDepth If set to true, accurate penetration depth is computed.
-      /// \return Collision proximity objectects for all colliding pairs of objects.
-      ///
-      std::vector<CollisionProxy> getCollisionDistance(bool self, bool computePenetrationDepth = true);
-
-      /**
-       * @brief      Gets the collision world links.
-       *
-       * @return     The collision world links.
-       */
-      std::vector<std::string> getCollisionWorldLinks()
-      {
-          std::vector<std::string> tmp;
-          for (fcl::CollisionObject* object : fcl_objects_)
-          {
-              KinematicElement* element = reinterpret_cast<KinematicElement*>(object->getUserData());
-              if(!element->ClosestRobotLink)
-              {
-                  tmp.push_back(element->Segment.getName());
-              }
-          }
-          return tmp;
-      }
-
-      /**
-       * @brief      Gets the collision robot links.
-       *
-       * @return     The collision robot links.
-       */
-      std::vector<std::string> getCollisionRobotLinks()
-      {
-          std::vector<std::string> tmp;
-          for (fcl::CollisionObject* object : fcl_objects_)
-          {
-              KinematicElement* element = reinterpret_cast<KinematicElement*>(object->getUserData());
-              if(element->ClosestRobotLink)
-              {
-                  tmp.push_back(element->Segment.getName());
-              }
-          }
-          return tmp;
-      }
-
-      Eigen::Vector3d getTranslation(const std::string & name);
-
-      inline void setACM(const AllowedCollisionMatrix& acm)
-      {
-          acm_ = acm;
-      }
-
-      ///
-      /// \brief Creates the collision scene from kinematic elements.
-      /// \param objects Vector kinematic element pointers of collision objects.
-      ///
-      void updateCollisionObjects(const std::map<std::string, std::shared_ptr<KinematicElement>>& objects);
-
-      ///
-      /// \brief Updates collision object transformations from the kinematic tree.
-      ///
-      void updateCollisionObjectTransforms();
-
-    private:
-
-     static std::shared_ptr<fcl::CollisionObject> constructFclCollisionObject(std::shared_ptr<KinematicElement> element);
-
-     std::map<std::string, std::shared_ptr<fcl::CollisionObject>> fcl_cache_;
-
-     std::vector<fcl::CollisionObject*> fcl_objects_;
-
-     /// Indicate if distance computation is required
-     bool compute_dist;
-
-     std::string root_name_;
-
-     /// The allowed collisiom matrix
-     AllowedCollisionMatrix acm_;
-  };
-
-  typedef std::shared_ptr<CollisionScene> CollisionScene_ptr;
-
   struct AttachedObject
   {
       AttachedObject() : Parent("") {}
