+++ exotica/src/Scene.cpp
@@ -31,382 +31,13 @@
  */
 
 #include "exotica/Scene.h"
+#include <exotica/Setup.h>
 #include <iostream>
 #include <fstream>
 #include <string>
-#include <fcl/BVH/BVH_model.h>
-#include <fcl/shape/geometric_shapes.h>
-#include <fcl/octree.h>
 
-namespace fcl_convert
-{
-  void fcl2Eigen(const fcl::Vec3f & fcl, Eigen::Vector3d & eigen)
-  {
-    eigen(0) = fcl.data.vs[0];
-    eigen(1) = fcl.data.vs[1];
-    eigen(2) = fcl.data.vs[2];
-  }
-
-  void fcl2Eigen(const fcl::Transform3f & fcl, Eigen::Vector3d & eigen)
-  {
-    eigen(0) = fcl.getTranslation().data.vs[0];
-    eigen(1) = fcl.getTranslation().data.vs[1];
-    eigen(2) = fcl.getTranslation().data.vs[2];
-  }
-
-  void fcl2EigenTranslation(const fcl::Vec3f & fcl, Eigen::Vector3d & eigen)
-  {
-    eigen(0) = fcl.data.vs[0];
-    eigen(1) = fcl.data.vs[1];
-    eigen(2) = fcl.data.vs[2];
-  }
-
-  fcl::Transform3f KDL2fcl(const KDL::Frame& frame)
-  {
-      return fcl::Transform3f(fcl::Matrix3f(frame.M.data[0], frame.M.data[1], frame.M.data[2], frame.M.data[3], frame.M.data[4], frame.M.data[5], frame.M.data[6], frame.M.data[7], frame.M.data[8]), fcl::Vec3f(frame.p.x(), frame.p.y(), frame.p.z()));
-  }
-}
 namespace exotica
 {
-
-AllowedCollisionMatrix::AllowedCollisionMatrix()
-{
-
-}
-
-AllowedCollisionMatrix::AllowedCollisionMatrix(const AllowedCollisionMatrix& acm)
-{
-    entries_ = acm.entries_;
-}
-
-void AllowedCollisionMatrix::clear()
-{
-    entries_.clear();
-}
-
-bool AllowedCollisionMatrix::hasEntry(const std::string& name) const
-{
-    return entries_.find(name)==entries_.end();
-}
-
-void AllowedCollisionMatrix::setEntry(const std::string& name1, const std::string& name2)
-{
-    entries_[name1].insert(name2);
-}
-
-void AllowedCollisionMatrix::getAllEntryNames(std::vector<std::string>& names) const
-{
-    names.clear();
-    for(auto& it : entries_)
-    {
-        names.push_back(it.first);
-    }
-}
-bool AllowedCollisionMatrix::getAllowedCollision(const std::string& name1, const std::string& name2) const
-{
-    auto it = entries_.find(name1);
-    if(it==entries_.end()) return true;
-    return it->second.find(name2)==it->second.end();
-}
-
-///////////////////////////////////////////////////////////////
-/////////////////////// Collision Scene ///////////////////////
-///////////////////////////////////////////////////////////////
-  CollisionScene::CollisionScene(const std::string& root_name)
-      : root_name_(root_name)
-  {
-  }
-
-  CollisionScene::~CollisionScene()
-  {
-  }
-
-  void CollisionScene::updateCollisionObjects(const std::map<std::string, std::shared_ptr<KinematicElement>>& objects)
-  {
-      fcl_objects_.resize(objects.size());
-      int i=0;
-      for(const auto& object : objects)
-      {
-          std::shared_ptr<fcl::CollisionObject> new_object;
-
-          const auto& cache_entry = fcl_cache_.find(object.first);
-          if(cache_entry == fcl_cache_.end())
-          {
-              new_object = constructFclCollisionObject(object.second);
-              fcl_cache_[object.first] = new_object;
-          }
-          else
-          {
-              new_object = cache_entry->second;
-          }
-          fcl_objects_[i++] = new_object.get();
-      }
-  }
-
-  void CollisionScene::updateCollisionObjectTransforms()
-  {
-      for(fcl::CollisionObject* collision_object : fcl_objects_)
-      {
-          KinematicElement* element = reinterpret_cast<KinematicElement*>(collision_object->getUserData());
-          collision_object->setTransform(fcl_convert::KDL2fcl(element->Frame));
-          collision_object->computeAABB();
-      }
-  }
-
-  // This function was copied from 'moveit_core/collision_detection_fcl/src/collision_common.cpp'
-  std::shared_ptr<fcl::CollisionObject> CollisionScene::constructFclCollisionObject(std::shared_ptr<KinematicElement> element)
-  {
-      // Maybe use cache here?
-
-      shapes::ShapeConstPtr shape = element->Shape;
-      boost::shared_ptr<fcl::CollisionGeometry> geometry;
-      if (shape->type == shapes::PLANE)  // shapes that directly produce CollisionGeometry
-      {
-        // handle cases individually
-        switch (shape->type)
-        {
-          case shapes::PLANE:
-          {
-            const shapes::Plane* p = static_cast<const shapes::Plane*>(shape.get());
-            geometry.reset(new fcl::Plane(p->a, p->b, p->c, p->d));
-          }
-          break;
-          default:
-            break;
-        }
-      }
-      else
-      {
-        switch (shape->type)
-        {
-          case shapes::SPHERE:
-          {
-            const shapes::Sphere* s = static_cast<const shapes::Sphere*>(shape.get());
-            geometry.reset(new fcl::Sphere(s->radius));
-          }
-          break;
-          case shapes::BOX:
-          {
-            const shapes::Box* s = static_cast<const shapes::Box*>(shape.get());
-            const double* size = s->size;
-            geometry.reset(new fcl::Box(size[0], size[1], size[2]));
-          }
-          break;
-          case shapes::CYLINDER:
-          {
-            const shapes::Cylinder* s = static_cast<const shapes::Cylinder*>(shape.get());
-            geometry.reset(new fcl::Cylinder(s->radius, s->length));
-          }
-          break;
-          case shapes::CONE:
-          {
-            const shapes::Cone* s = static_cast<const shapes::Cone*>(shape.get());
-            geometry.reset(new fcl::Cone(s->radius, s->length));
-          }
-          break;
-          case shapes::MESH:
-          {
-            fcl::BVHModel<fcl::OBBRSS>* g = new fcl::BVHModel<fcl::OBBRSS>();
-            const shapes::Mesh* mesh = static_cast<const shapes::Mesh*>(shape.get());
-            if (mesh->vertex_count > 0 && mesh->triangle_count > 0)
-            {
-              std::vector<fcl::Triangle> tri_indices(mesh->triangle_count);
-              for (unsigned int i = 0; i < mesh->triangle_count; ++i)
-                tri_indices[i] =
-                    fcl::Triangle(mesh->triangles[3 * i], mesh->triangles[3 * i + 1], mesh->triangles[3 * i + 2]);
-
-              std::vector<fcl::Vec3f> points(mesh->vertex_count);
-              for (unsigned int i = 0; i < mesh->vertex_count; ++i)
-                points[i] = fcl::Vec3f(mesh->vertices[3 * i], mesh->vertices[3 * i + 1], mesh->vertices[3 * i + 2]);
-
-              g->beginModel();
-              g->addSubModel(points, tri_indices);
-              g->endModel();
-            }
-            geometry.reset(g);
-          }
-          break;
-          case shapes::OCTREE:
-          {
-            const shapes::OcTree* g = static_cast<const shapes::OcTree*>(shape.get());
-            geometry.reset(new fcl::OcTree(g->octree));
-          }
-          break;
-          default:
-            throw_pretty("This shape type ("<<((int)shape->type)<<") is not supported using FCL yet");
-        }
-      }
-      geometry->computeLocalAABB();
-      geometry->setUserData(reinterpret_cast<void*>(element.get()));
-      std::shared_ptr<fcl::CollisionObject> ret(new fcl::CollisionObject(geometry));
-      ret->setUserData(reinterpret_cast<void*>(element.get()));
-
-      return ret;
-  }
-
-  bool CollisionScene::isAllowedToCollide(fcl::CollisionObject* o1, fcl::CollisionObject* o2, bool self, CollisionScene* scene)
-  {
-      KinematicElement* e1 = reinterpret_cast<KinematicElement*>(o1->getUserData());
-      KinematicElement* e2 = reinterpret_cast<KinematicElement*>(o2->getUserData());
-
-      bool isRobot1 = e1->isRobotLink || e1->ClosestRobotLink;
-      bool isRobot2 = e2->isRobotLink || e2->ClosestRobotLink;
-      // Don't check collisions between world objects
-      if(!isRobot1 && !isRobot2) return false;
-      // Skip self collisions if requested
-      if(isRobot1 && isRobot2 && !self) return false;
-      // Skip collisions between shapes within the same objects
-      if(e1->Parent==e2->Parent) return false;
-      // Skip collisions between bodies attached to the same object
-      if(e1->ClosestRobotLink&&e2->ClosestRobotLink&&e1->ClosestRobotLink==e2->ClosestRobotLink) return false;
-
-      if(isRobot1 && isRobot2)
-      {
-          const std::string& name1 = e1->ClosestRobotLink?e1->ClosestRobotLink->Segment.getName():e1->Parent->Segment.getName();
-          const std::string& name2 = e2->ClosestRobotLink?e2->ClosestRobotLink->Segment.getName():e2->Parent->Segment.getName();
-          return scene->acm_.getAllowedCollision(name1, name2);
-      }
-      return true;
-  }
-
-  bool CollisionScene::collisionCallback(fcl::CollisionObject* o1, fcl::CollisionObject* o2, void* data)
-  {
-      CollisionData* data_ = reinterpret_cast<CollisionData*>(data);
-
-      if(!isAllowedToCollide(o1, o2, data_->Self, data_->Scene)) return false;
-
-      data_->Request.num_max_contacts = 1000;
-      data_->Result.clear();
-      fcl::collide(o1,o2,data_->Request, data_->Result);
-      data_->Done = data_->Result.isCollision();
-      return data_->Done;
-  }
-
-  bool CollisionScene::collisionCallbackDistance(fcl::CollisionObject* o1, fcl::CollisionObject* o2, void* data, double& dist)
-  {
-      DistanceData* data_ = reinterpret_cast<DistanceData*>(data);
-
-      if(!isAllowedToCollide(o1, o2, data_->Self, data_->Scene)) return false;
-      data_->Request.enable_nearest_points = false;
-      data_->Result.clear();
-      fcl::distance(o1,o2,data_->Request, data_->Result);
-      int num_contacts = data_->Distance = std::min(data_->Distance, data_->Result.min_distance);
-
-      CollisionProxy p;
-      p.e1 = reinterpret_cast<KinematicElement*>(o1->getUserData());
-      p.e2 = reinterpret_cast<KinematicElement*>(o2->getUserData());
-
-      p.distance = data_->Result.min_distance;
-      fcl_convert::fcl2Eigen(data_->Result.nearest_points[0], p.contact1);
-      fcl_convert::fcl2Eigen(data_->Result.nearest_points[1], p.contact2);
-      p.normal1 = p.contact1.normalized();
-      p.normal2 = p.contact2.normalized();
-      p.distance = (p.contact1-p.contact2).norm();
-      data_->Distance = std::min(data_->Distance, p.distance);
-      data_->Proxies.push_back(p);
-
-      return false;
-  }
-
-  bool CollisionScene::collisionCallbackContacts(fcl::CollisionObject* o1, fcl::CollisionObject* o2, void* data, double& dist)
-  {
-      ContactData* data_ = reinterpret_cast<ContactData*>(data);
-
-      if(!isAllowedToCollide(o1, o2, data_->Self, data_->Scene)) return false;
-
-      data_->Request.enable_contact = true;
-      data_->Request.num_max_contacts = 1000;
-      data_->Result.clear();
-      int num_contacts = fcl::collide(o1, o2, data_->Request, data_->Result);
-      CollisionProxy p;
-      p.e1 = reinterpret_cast<KinematicElement*>(o1->getUserData());
-      p.e2 = reinterpret_cast<KinematicElement*>(o2->getUserData());
-
-      if(num_contacts>0)
-      {
-          p.distance = -data_->Result.getContact(0).penetration_depth;
-          for(int i=0; i<num_contacts; i++)
-          {
-              const fcl::Contact& contact = data_->Result.getContact(i);
-              if(p.distance>-contact.penetration_depth)
-              {
-                  p.distance=-contact.penetration_depth;
-                  if(reinterpret_cast<KinematicElement*>(contact.o1->getUserData())==p.e1)
-                  {
-                      fcl_convert::fcl2Eigen(contact.pos, p.contact1);
-                      fcl_convert::fcl2Eigen(contact.normal, p.normal1);
-                      p.normal1.normalize();
-                  }
-                  else if(reinterpret_cast<KinematicElement*>(contact.o2->getUserData())==p.e2)
-                  {
-                      fcl_convert::fcl2Eigen(contact.pos, p.contact2);
-                      fcl_convert::fcl2Eigen(contact.normal, p.normal2);
-                      p.normal2.normalize();
-                  }
-              }
-          }
-      }
-      else
-      {
-          data_->DistanceRequest.enable_nearest_points = true;
-          data_->DistanceRequest.gjk_solver_type = fcl::GJKSolverType::GST_LIBCCD;
-          fcl::distance(o1,o2,data_->DistanceRequest, data_->DistanceResult);
-          fcl_convert::fcl2Eigen(data_->DistanceResult.nearest_points[0], p.contact1);
-          fcl_convert::fcl2Eigen(data_->DistanceResult.nearest_points[1], p.contact2);
-          p.normal1 = p.contact1.normalized();
-          p.normal2 = p.contact2.normalized();
-          p.distance = (p.contact1-p.contact2).norm();
-      }
-      data_->Distance = std::min(data_->Distance, p.distance);
-      data_->Proxies.push_back(p);
-
-      return false;
-  }
-
-  bool CollisionScene::isStateValid(bool self)
-  {
-      std::shared_ptr<fcl::BroadPhaseCollisionManager> manager(new fcl::DynamicAABBTreeCollisionManager());
-      manager->registerObjects(fcl_objects_);
-      CollisionData data(this);
-      data.Self = self;
-      manager->collide(&data, &CollisionScene::collisionCallback);
-      return !data.Result.isCollision();
-  }
-
-  std::vector<CollisionProxy> CollisionScene::getCollisionDistance(bool self, bool computePenetrationDepth)
-  {
-      std::shared_ptr<fcl::BroadPhaseCollisionManager> manager(new fcl::DynamicAABBTreeCollisionManager());
-      manager->registerObjects(fcl_objects_);
-      if(computePenetrationDepth)
-      {
-          ContactData data(this);
-          data.Self = self;
-          manager->distance(&data, &CollisionScene::collisionCallbackContacts);
-          return data.Proxies;
-      }
-      else
-      {
-          DistanceData data(this);
-          data.Self = self;
-          manager->distance(&data, &CollisionScene::collisionCallbackDistance);
-          return data.Proxies;
-      }
-  }
-
-
-  Eigen::Vector3d CollisionScene::getTranslation(const std::string & name)
-  {
-      for(fcl::CollisionObject* object : fcl_objects_)
-      {
-          KinematicElement* element = reinterpret_cast<KinematicElement*>(object->getUserData());
-          if(element->Segment.getName()==name)
-          {
-              return Eigen::Map<Eigen::Vector3d>(element->Frame.p.data);
-          }
-      }
-      throw_pretty("Robot not found!");;
-  }
-
 ///////////////////////////////////////////////////////////////
 /////////////////////// EXOTica Scene ///////////////////////
 ///////////////////////////////////////////////////////////////
@@ -467,7 +98,7 @@ bool AllowedCollisionMatrix::getAllowedCollision(const std::string& name1, const
                   << "/PlanningScene'");
       }
 
-      collision_scene_.reset(new CollisionScene(kinematica_.getRootFrameName()));
+      collision_scene_ = Setup::createCollisionScene("CollisionSceneFCL");
       collision_scene_->updateCollisionObjects(kinematica_.getCollisionTreeMap());
 
       AllowedCollisionMatrix acm;
